Undo test changes
